--- a/app.py
+++ b/app.py
@@ -1,6 +1,7 @@
 import time
+import asyncpg
 
-from quart import Quart, jsonify, request
+from quart import Quart, jsonify, request, current_app
 from db import search
 
 app = Quart(__name__)
@@ -12,10 +13,16 @@ async def index():
     if not query:
         return jsonify({})
     start = time.time()
-    res = search(query)
+    async with current_app.pool.acquire() as connection:
+        res = await search(query, connection)
     return jsonify({
         'time': time.time() - start,
-        'data': res
+        'data': [{k: v for (k,v) in r.items()} for r in res]
     })
+
+@app.before_first_request
+async def create_db():
+    dsn = 'postgres://postgres:postgres@localhost:5432/datagouvfr'
+    app.pool = await asyncpg.create_pool(dsn, max_size=20)
 
 app.run()