--- a/requirements.txt
+++ b/requirements.txt
@@ -1,3 +1,3 @@
 Quart
-psycopg2-binary
+asynpg
 csvkit